third_party/WebKit/LayoutTests/editing/execCommand/overtype-support.html

<!DOCTYPE html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description('This test checks that the Overwrite mode is: 1) not exposed to JavaScript 2) queryCommandState returns false.');

shouldBeFalse("document.queryCommandSupported('OverWrite')");
shouldBeFalse("document.queryCommandState('OverWrite')");

var successfullyParsed = true;
</script>
</body>
</html>
